+++ library/src/validations/ulid/ulid.ts
@@ -2,7 +2,7 @@ import type { ErrorMessage, PipeResult } from '../../types.ts';
 import { getOutput, getPipeIssues } from '../../utils/index.ts';
 
 /**
- * Creates a validation functions that validates a [ULID](https://github.com/ulid/spec).
+ * Creates a validation function that validates a [ULID](https://github.com/ulid/spec).
  *
  * @param error The error message.
  *